+++ test/utils/dom-helpers.js
